Refactors supposed To Make It More Extensible
I rewrite the library based on what I learned from trying to extend the
original version. This rewrite is backward compatible, and introduces
additional features, including the ability to subscribe to all events,
and configure multiple reporters (i.e. to print to screen, and to files,
or a database).

The tests are printed as they complete, and TAP support is improved
(fixed some bugs like error stacks messing up YAML indentation). There
is no need for this library to bail when tests are broken because the
library reports broken as a type of failure. The examples for using bail
out on TAPs website are more appropriate for skipping, and failing
unless the entire suite is skipped or fails at which point the chainable
features of this library are betters suited. For instance you can run a
test to verify the OS state, and throw if the tests should bail. The
suite would be put in the `then` in this case, and would only be
executed if the first test passes.

Reporters are much simpler to write now because the Tally reporter and
event pubsub take care of the need to keep track of test outcomes.

A preliminary index.d.ts file is introduced to start adding TypeScript
support.