Removes Unused Dependencies and Simplifies Arg Processing (#12)
At one point I performed validation while reading envvars. As I made
changes so the envvars could be established programmatically, that
validation moved. This removes the artifacts from the previous way that
I validated args and envvars.

Also simplifies arg processing. There was an unnecessary condition in 
`findMatch` that was also an artifact of a previous way that I read 
arguments. This condition is removed by aligning the property names 
of `Switch` and `Swatch`.