Adds Test Log Output to Markdown Formatter
When writing tests to produce documentation, especially for proofs of
concept, printing the log output to markdown can save a lot of time, and
make our tests a bit more like a live notebook. This adds log output to
the markdown formatter, and updates the error formatting for markdown,
so they both follow the same convention.