fix: fixes unexpected test passing
Similar to passing arguments in to a function in the wrong order, with
currying, developers might pass functions in the wrong order.

Given the following `then` curry:

```javascript
'it should': (err, actual) => () => {}
```

The test was not being executed because the test runner assumed the
first function expected 3 arguments (expect, err, actual) and that it
was the only function.

This commit changes the logic so it always checks for curried functions,
regardless of the number of arguments being passed.

I think this is a hotfix and that it isn't a breaking change given the
assumptions in the existing code and the sparse documentation about that
feature (in the TS tests and TS README examples).