feat: adds support for es modules
I used require to load modules in the runner, but that only has support
for commonJS modules. There is a dynamic import global that is
functional like require, but is asynchronous and works with both
commonJS and ES modules. This adds tests to verify that the change to
using that works. It also introduces mjs and cjs to the glob match for
test files.